backend/requirements.txt
pytest===7.2.1
gunicorn===20.1.0
Flask-Migrate===4.0.4
python-dotenv===1.0.0
psycopg2-binary===2.9.5
Flask-SQLAlchemy===3.0.3